added modular axe & pickaxe

+++ src/main/java/se/mickelus/tetra/module/SynergyData.java
@@ -4,6 +4,6 @@ public class SynergyData {
     public String[] moduleVariants = new String[0];
 
     public CapabilityData capabilities = new CapabilityData();
-    public float damage = 0;
-    public float speed = 0;
+    public double damage = 0;
+    public double speed = 0;
 }
